+++ packages/utilities/index.ts
@@ -12,6 +12,7 @@
 export * from "./src/doc/IDaemonRequest";
 export * from "./src/doc/IDaemonResponse";
 export * from "./src/DaemonRequest";
+export * from "./src/ExecUtils";
 export * from "./src/ImperativeConfig";
 export * from "./src/JSONUtils";
 export * from "./src/JsUtils";